+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Tideland Go Library
 
+## 2016-10-xx
+
+- Backend in *monitoring* is now lazy loading
+
 ## 2016-10-13
 
 - *Etc* now can handle templates substituted with values from other